Extract shared CASA value and table-read crates
