+++ Cargo.toml
@@ -5,8 +5,10 @@ members = [
     "crates/casa-calibration",
     "crates/casars-imagebrowser-protocol",
     "crates/casars-tablebrowser-protocol",
+    "crates/casa-values",
     "crates/casa-types",
     "crates/casa-aipsio",
+    "crates/casa-table-read",
     "crates/casa-measures-tools",
     "crates/casa-tables",
     "crates/casa-lattices",
