+++ crates/casa-measures-data/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 update = ["dep:ureq"]
 
 [dependencies]
+casa-table-read = { path = "../casa-table-read" }
 flate2 = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
