+++ crates/casa-table-read/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "casa-aipsio"
+name = "casa-table-read"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
-description = "AipsIO binary serialization codec (internal)"
+description = "Minimal read-only CASA table reader (internal)"
 repository.workspace = true
 publish = false
 
@@ -12,6 +12,4 @@ publish = false
 workspace = true
 
 [dependencies]
-casa-values = { path = "../casa-values" }
-ndarray.workspace = true
 thiserror.workspace = true
