+++ crates/casa-table-read/src/aipsio_buf.rs
