+++ crates/casa-table-read/src/incremental_stman.rs
